// ps-angular-forms/src/app/model/user-settings.ts
export class UserSettings{
    emailAddress: string = '';
    username: string = '';
    firstName: string = '';
    lastName: string = '';
    city: string = '';
    state: string = '';
    zip: string = '';
    membershipType: string = '';
    weeklyNewsletter: boolean = false;
    dailySpecials: boolean = false;
    clearanceSales: boolean = false;
    userInterfaceStyle: string = '';
    notes: string = '';
}